api/requirements.txt
Django==4.1
djangorestframework==3.13.1
djangorestframework-simplejwt==5.2.0
python-decouple==3.6
cloudinary==1.30.0
django-cors-headers==3.13.0
psycopg2-binary==2.9.6
gunicorn==20.1.0
whitenoise==6.2.0